Add www.erebot.net & MultiProjectBuildFactory.
This makes buildbot manage www.erebot.net by reusing the existing Live
scheduler and builder, except that it uses a different set of steps than
for Erebot's Live builder. This is accomplished by adding a level of
indirection in the builder's factory, calling the specific BuildFactory
for a project rather than a unique (shared) factory.

This also removes all tests and API doc generation for Erebot_API because
those are already covered by Erebot.