Move locks to dedicated file & remove live builders.
Move builder locks to the file dedicated to locks.
Remove both the "live" and "live_www" factories and their associated builder
as this task is better handled through a deployment tool such as puppet.
Remove the repository for www.erebot.net from the conf altogether.
Simplify some of the config.